Fixed a massive number of trivial GCC warnings.
Most of them are about vararg-style formatting.
It is really great to have these warnings, but they generate too much trash which had to be cleaned.

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@10568 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ cm/CollisionModel_load.cpp
@@ -3264,13 +3264,13 @@ idCollisionModelManagerLocal::PrintModelInfo
 ==================
 */
 void idCollisionModelManagerLocal::PrintModelInfo( const cm_model_t *model ) {
-	common->Printf( "%6i vertices (%i KB)\n", model->numVertices, (model->numVertices * sizeof(cm_vertex_t))>>10 );
-	common->Printf( "%6i edges (%i KB)\n", model->numEdges, (model->numEdges * sizeof(cm_edge_t))>>10 );
+	common->Printf( "%6i vertices (%lli KB)\n", model->numVertices, int64(model->numVertices * sizeof(cm_vertex_t))>>10 );
+	common->Printf( "%6i edges (%lli KB)\n", model->numEdges, int64(model->numEdges * sizeof(cm_edge_t))>>10 );
 	common->Printf( "%6i polygons (%i KB)\n", model->numPolygons, model->polygonMemory>>10 );
 	common->Printf( "%6i brushes (%i KB)\n", model->numBrushes, model->brushMemory>>10 );
-	common->Printf( "%6i nodes (%i KB)\n", model->numNodes, (model->numNodes * sizeof(cm_node_t))>>10 );
-	common->Printf( "%6i polygon refs (%i KB)\n", model->numPolygonRefs, (model->numPolygonRefs * sizeof(cm_polygonRef_t))>>10 );
-	common->Printf( "%6i brush refs (%i KB)\n", model->numBrushRefs, (model->numBrushRefs * sizeof(cm_brushRef_t))>>10 );
+	common->Printf( "%6i nodes (%lli KB)\n", model->numNodes, int64(model->numNodes * sizeof(cm_node_t))>>10 );
+	common->Printf( "%6i polygon refs (%lli KB)\n", model->numPolygonRefs, int64(model->numPolygonRefs * sizeof(cm_polygonRef_t))>>10 );
+	common->Printf( "%6i brush refs (%lli KB)\n", model->numBrushRefs, int64(model->numBrushRefs * sizeof(cm_brushRef_t))>>10 );
 	common->Printf( "%6i internal edges\n", model->numInternalEdges );
 	common->Printf( "%6i sharp edges\n", model->numSharpEdges );
 	common->Printf( "%6i contained polygons removed\n", model->numRemovedPolys );
